算法优先文法解析复杂多项式完成

+++ parse_polynomial/parse_poly/parse_poly/global_val.cpp
@@ -37,6 +37,6 @@ int cmp_prio[10][10]={
 
 //std::regex pattern_num("^(-?\\d*|-?\\d*.\\d*)$");
 //std::regex pattern_ops("^N(+|-|*|/|\\^)N$");
-std::stack<int> N_stack;
+std::stack<float> N_stack;
 std::map<char,int> sigmap;
