+++ parse_polynomial/parse_poly/parse_poly/parse.h
@@ -5,9 +5,7 @@
 #include<string>
 
 
-extern int parse_poly(std::string expr);
-
-
+extern float parse_poly(std::string expr);
 
 
 #endif
